% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dijkstra.R
\name{dijkstra}
\alias{dijkstra}
\title{Dijkstra algorithm}
\source{
\href{https://en.wikipedia.org/wiki/Dijkstra\%27s_algorithm}{Wikipedia}
}
\usage{
dijkstra(graph, init_node)
}
\arguments{
\item{graph}{a data.frame}

\item{init_node}{a number}
}
\value{
a vector representing the shortest path to every other node from the starting node
}
\description{
\code{dijkstra} returns the greatest common divisor of two integers.
}
\details{
The algorithm takes a graph and an initial node and calculates the shortest path from the inital node to every other node in the graph
}
