% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euclidean.R
\name{euclidean}
\alias{euclidean}
\title{Euclidean algorithm}
\usage{
euclidean(x, y)
}
\arguments{
\item{x}{a number or integer.}

\item{y}{a number or integer.}
}
\value{
If the two inputs are integer or numeric, then the output
  will be an integer, specifically the highest number that will divide the two inputs. Otherwise the function will stop and show an error stating the inputs need to be an integer or numeric. 

  See
  \url{https://en.wikipedia.org/wiki/Euclidean_algorithm} for more details.
}
\description{
\code{euclidean} returns the greatest common divisor of two integers.
}
\details{
The algorithm is a method for computing the largest possible number that divides two numbers without a remainder. It works by replacing the larger number by its difference with the smaller of the two numbers and repeating this process until the two numbers become equal and the remainder is zero.
}
