% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-lab3.R
\docType{package}
\name{lab3}
\alias{lab3}
\title{lab3: A package for computing the Euclidean and Dijkstras values}
\description{
The lab3 package provides two important functions:
euclidean and dijkstras.
}
