% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wiki_graph}
\alias{wiki_graph}
\title{wiki_graph}
\format{
A data frame with columns:
\describe{
 \item{v1}{Starting node.}
 \item{v2}{Ending node.}
 \item{w}{Weight/distance from v1 node to v2 node.}
}
}
\source{

}
\usage{
data(wiki_graph)
}
\description{
dataset for the dijkstra function
}
\details{
Data for dikstra function
}
\keyword{datasets}
