% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euclidean.R
\name{euclidean}
\alias{euclidean}
\title{Euclidean Algorithm}
\usage{
euclidean(a, b)
}
\arguments{
\item{a}{Numeric Number}

\item{b}{Numeric Number}
}
\value{
The Greatest Common Divisor of two numbers.
}
\description{
Find Greatest Common Divisor(GCD) of Two Given Parameters
}
\examples{
euclidean(123612, 13892347912)
euclidean(100, 1000)

}
\references{
https://en.wikipedia.org/wiki/Euclidean_algorithm
}
