% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lab3.R
\docType{package}
\name{lab3}
\alias{lab3}
\title{Lab3 Package}
\description{
Contains two functions for implementation of Euclidean and Dijkstra's  algorithms
}
\author{
SILA KILICOGLU (silki753@student.liu.se) and SAMI FURKAN YILDIRIM (samyi572@student.liu.se)
}
